Initial implementation of auto-updater
