+++ io_scene_nif/update.py
@@ -38,41 +38,40 @@
 # ***** END LICENSE BLOCK *****
 
 import bpy
-from io_scene_nif.utils.updater import addon_updater_ops
+from io_scene_nif import addon_updater_ops
 
-@addon_updater_ops.make_annotations
+
+# @addon_updater_ops.make_annotations
 class UpdaterPreferences(bpy.types.AddonPreferences):
-    """Demo bare-bones preferences"""
     bl_idname = __package__
 
     # addon updater preferences
-
-    auto_check_update = bpy.props.BoolProperty(
+    auto_check_update: bpy.props.BoolProperty(
         name="Auto-check for Update",
         description="If enabled, auto-check for updates using an interval",
         default=False,
     )
-    updater_intrval_months = bpy.props.IntProperty(
+    updater_interval_months: bpy.props.IntProperty(
         name='Months',
         description="Number of months between checking for updates",
         default=0,
         min=0
     )
-    updater_intrval_days = bpy.props.IntProperty(
+    updater_interval_days: bpy.props.IntProperty(
         name='Days',
         description="Number of days between checking for updates",
         default=7,
         min=0,
         max=31
     )
-    updater_intrval_hours = bpy.props.IntProperty(
+    updater_interval_hours: bpy.props.IntProperty(
         name='Hours',
         description="Number of hours between checking for updates",
         default=0,
         min=0,
         max=23
     )
-    updater_intrval_minutes = bpy.props.IntProperty(
+    updater_interval_minutes: bpy.props.IntProperty(
         name='Minutes',
         description="Number of minutes between checking for updates",
         default=0,
@@ -89,7 +88,7 @@ class UpdaterPreferences(bpy.types.AddonPreferences):
 
         # updater draw function
         # could also pass in col as third arg
-        addon_updater_ops.update_settings_ui(self, context)
+        addon_updater_ops.update_settings_ui(self, context, col)
 
     # Alternate draw function, which is more condensed and can be
     # placed within an existing draw function. Only contains:
@@ -101,4 +100,5 @@ class UpdaterPreferences(bpy.types.AddonPreferences):
     # col = mainrow.column()
     # col.scale_y = 2
     # col.operator("wm.url_open","Open webpage ").url=addon_updater_ops.updater.website
+
